Add Join contest Screen
