+++ src/Routes/index.js
@@ -6,7 +6,7 @@ import HomePage from '../components/Home/Home';
 import Login from '../components/Auth/Login';
 import SignUp from '../components/Auth/SignUp';
 import Contests from '../components/Contests/Contests';
-import CreateContest from '../components/Contests/CreateContest/CreateContest';
+import JoinContest from '../components/Contests/JoinContest/JoinContest';
 
 export default () => (
     <Router history={history}>
@@ -15,7 +15,7 @@ export default () => (
             <Route path="/login" component={Login} />
             <Route path="/register" component={SignUp} />
             <Route path="/contests" component={Contests} />
-            <Route path="/create-contest" component={CreateContest} />
+            <Route path="/join-contest" component={JoinContest} />
         </Switch>
     </Router>
 );
